remove unintended white space on mobile
when viewing the webstore on mobile, or inspecting from the browser using
a mobile device, there was extra white space alone the right side of
the page. turns out that the ItemGroup component was causing it.

we're using a bootstrap grid class that was adding x/y gutters in between
each item. this commit removes the horizontal (x) gutter, opting only
for the vertical (y) gutter. the page still looks good on desktop too.

- ref: https://getbootstrap.com/docs/5.0/layout/gutters/
- ref: #211